Retrieving and Updating Variables
This commit addresses issues with retrieving and updating variables. Additionally, it adds the motion control boxes as a floating window to control all 3 axis of motion.

+++ src/library_SPIIMotionController/status/status_position.cpp
@@ -29,7 +29,7 @@ MotorAxis Status_PositionPerAxis::getAxis() const
     return this->currentAxis;
 }
 
-int Status_PositionPerAxis::getPosition() const
+double Status_PositionPerAxis::getPosition() const
 {
     return position;
 }
